# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/dcp/tasks:
    get:
      operationId: listDCPTasks
      parameters:
      - description: Optional project filter; only dcp-lab is accepted.
        in: query
        name: project
        schema:
          description: Optional project filter; only dcp-lab is accepted.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListDCPTasksResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List durable synthetic DCP tasks
      tags:
      - dcpTasks
    post:
      operationId: submitDCPTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitDCPTaskRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DCPTaskResponse'
          description: OK
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DCPTaskResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Submit one model-free synthetic DCP task in SUBMITTED
      tags:
      - dcpTasks
  /api/v1/dcp/tasks/{taskId}:
    get:
      operationId: getDCPTask
      parameters:
      - description: Durable DCP task identifier.
        in: path
        name: taskId
        required: true
        schema:
          description: Durable DCP task identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DCPTaskResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Read one durable synthetic DCP task
      tags:
      - dcpTasks
  /api/v1/dcp/tasks/{taskId}/events:
    get:
      operationId: listDCPTaskEvents
      parameters:
      - description: Durable DCP task identifier.
        in: path
        name: taskId
        required: true
        schema:
          description: Durable DCP task identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListDCPTaskEventsResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List the monotonic event stream for one DCP task
      tags:
      - dcpTasks
  /api/v1/dcp/tasks/policy:
    post:
      operationId: submitDCPReviewLabPolicyTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControllersSubmitDCPPolicyTaskRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllersDCPPolicyTaskResponse'
          description: OK
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllersDCPPolicyTaskResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Submit one durable future DCP review-lab policy task
      tags:
      - dcpTasks
  /api/v1/dcp/v2/tasks:
    post:
      operationId: submitDCPV2TwinTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControllersSubmitDCPV2TwinTaskRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dcpv2TwinSubmitResult'
          description: OK
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dcpv2TwinSubmitResult'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Submit the one authorized DCP v2 integration-twin task
      tags:
      - dcpTasks
  /api/v1/dcp/v2/tasks/{taskId}:
    get:
      operationId: getDCPV2TwinTask
      parameters:
      - description: Durable DCP task identifier.
        in: path
        name: taskId
        required: true
        schema:
          description: Durable DCP task identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dcpv2TwinSnapshot'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Read the exact DCP v2 integration-twin lifecycle
      tags:
      - dcpTasks
  /api/v1/dcp/v2/tasks/{taskId}/release-event:
    post:
      operationId: wakeDCPV2TwinRelease
      parameters:
      - description: Durable DCP task identifier.
        in: path
        name: taskId
        required: true
        schema:
          description: Durable DCP task identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControllersWakeDCPV2TwinReleaseRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dcpv2TwinSnapshot'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Apply one immutable repository Release Train completion event
      tags:
      - dcpTasks
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    ControllersDCPPolicyTaskResponse:
      properties:
        duplicate:
          type: boolean
        task:
          $ref: '#/components/schemas/DomainDCPReviewLabPolicyTask'
      required:
      - task
      - duplicate
      type: object
...
    ControllersSubmitDCPPolicyTaskRequest:
      properties:
        profile:
          enum:
          - synthetic-pr
          - repo-only
          - live-runtime
          type: string
        prompt:
          maxLength: 512
          minLength: 1
          type: string
        repository:
          enum:
          - orenvlad-ai/dcp-review-lab
          - orenvlad-ai/wb-price-extension
          - orenvlad-ai/wb-browser-extension
          - orenvlad-ai/wb-core
          - orenvlad-ai/dcp-wbc-integration-lab
          type: string
        target:
          enum:
          - dcp-review-lab
          - wb-price-extension
          - wb-browser-extension
          - wb-core
          - dcp-wbc-integration-lab
          type: string
        taskId:
          maxLength: 64
          minLength: 1
          type: string
      required:
      - taskId
      - target
      - profile
      - repository
      - prompt
      type: object
    ControllersSubmitDCPV2TwinTaskRequest:
      properties:
        prompt:
          maxLength: 512
          minLength: 1
          type: string
        taskId:
          maxLength: 64
          minLength: 1
          type: string
      required:
      - taskId
      - prompt
      type: object
    ControllersWakeDCPV2TwinReleaseRequest:
      properties:
        deliveryId:
          maxLength: 256
          minLength: 1
          type: string
        payloadDigest:
          maxLength: 64
          minLength: 64
          type: string
        runId:
          format: int64
          minimum: 1
          type: integer
      required:
      - deliveryId
      - runId
      - payloadDigest
      type: object
    DCPApprovedScope:
      properties:
        schemaVersion:
          enum:
          - dcp.scope/v1
          type: string
        statement:
          type: string
      required:
      - schemaVersion
      - statement
      type: object
    DCPApprovedTask:
      properties:
        description:
          type: string
        schemaVersion:
          enum:
          - dcp.task/v1
          type: string
        title:
          type: string
      required:
      - schemaVersion
      - title
      - description
      type: object
    DCPRepositoryIdentity:
      properties:
        headSha:
          type: string
        identityDigest:
          type: string
        markerDigest:
          type: string
        path:
          type: string
        projectId:
          enum:
          - dcp-lab
          type: string
        repository:
          enum:
          - dcp-lab
          type: string
        schemaVersion:
          enum:
          - dcp.repository/v1
          type: string
      required:
      - schemaVersion
      - projectId
      - repository
      - path
      - headSha
      - markerDigest
      - identityDigest
      type: object
    DCPTask:
      properties:
        approvedDigest:
          type: string
        approvedScope:
          $ref: '#/components/schemas/DCPApprovedScope'
        approvedTask:
          $ref: '#/components/schemas/DCPApprovedTask'
        createdAt:
          format: date-time
          type: string
        idempotencyKey:
          type: string
        revision:
          format: int64
          type: integer
        state:
          enum:
          - SUBMITTED
          type: string
        target:
          $ref: '#/components/schemas/DCPRepositoryIdentity'
        taskId:
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - taskId
      - idempotencyKey
      - approvedTask
      - approvedScope
      - approvedDigest
      - target
      - state
      - revision
      - createdAt
      - updatedAt
      type: object
    DCPTaskEventResponse:
      properties:
        causationId:
          type: string
        correlationId:
          type: string
        eventId:
          type: string
        eventType:
          enum:
          - task.submitted
          - system.reconciled
          type: string
        evidenceDigest:
          type: string
        fromState:
          enum:
          - SUBMITTED
          type: string
        idempotencyKey:
          type: string
        integrityDigest:
          type: string
        occurredAt:
          format: date-time
          type: string
        payload: {}
        recordedAt:
          format: date-time
          type: string
        schemaVersion:
          enum:
          - dcp.event/v1
          type: string
        sequence:
          format: int64
          minimum: 1
          type: integer
        sourceId:
          type: string
        sourceKind:
          enum:
          - daemon
          type: string
        taskId:
          type: string
        taskRevision:
          format: int64
          minimum: 1
          type: integer
        toState:
          enum:
          - SUBMITTED
          type: string
      required:
      - taskId
      - sequence
      - eventId
      - schemaVersion
      - eventType
      - sourceKind
      - sourceId
      - correlationId
      - idempotencyKey
      - toState
      - taskRevision
      - occurredAt
      - recordedAt
      - payload
      - evidenceDigest
      - integrityDigest
      type: object
    DCPTaskResponse:
      properties:
        duplicate:
          type: boolean
        task:
          $ref: '#/components/schemas/DCPTask'
      required:
      - task
      - duplicate
      type: object
    Dcpv2TwinSnapshot:
      properties:
        actions:
          items:
            $ref: '#/components/schemas/DomainDCPV2Action'
          type: array
        admissions:
          items:
            $ref: '#/components/schemas/DomainDCPV2Admission'
          type: array
        commands:
          items:
            $ref: '#/components/schemas/DomainDCPV2Command'
          type: array
        events:
          items:
            $ref: '#/components/schemas/DomainDCPV2ExternalEvent'
          type: array
        incidents:
          items:
            $ref: '#/components/schemas/DomainDCPV2Incident'
          type: array
        native:
          $ref: '#/components/schemas/DomainDCPReviewLabPolicyTask'
        projection:
          $ref: '#/components/schemas/DomainDCPV2LifecycleProjection'
        results:
          items:
            $ref: '#/components/schemas/DomainDCPV2Result'
          type: array
        revisions:
          items:
            $ref: '#/components/schemas/DomainDCPV2Revision'
          type: array
        task:
          $ref: '#/components/schemas/DomainDCPV2Task'
      required:
      - task
      - native
      - revisions
      - commands
      - actions
      - admissions
      - results
      - incidents
      - events
      - projection
      type: object
    Dcpv2TwinSubmitResult:
      properties:
        duplicate:
          type: boolean
        native:
          $ref: '#/components/schemas/DomainDCPReviewLabPolicyTask'
        projection:
          $ref: '#/components/schemas/DomainDCPV2LifecycleProjection'
        task:
          $ref: '#/components/schemas/DomainDCPV2Task'
      required:
      - task
      - native
      - projection
      - duplicate
      type: object
...
    DomainDCPReviewLabPolicyTask:
      properties:
        admissionId:
          type: string
        cardNumber:
          format: int64
          type: integer
        createdAt:
          format: date-time
          type: string
        currentHeadSha:
          type: string
        errorCode:
          type: string
        incidentPacket:
          type: string
        mergeCommitSha:
          type: string
        payloadDigest:
          type: string
        policyVersion:
          type: string
        prNumber:
          format: int64
          type: integer
        prUrl:
          type: string
        previousHeadSha:
          type: string
        profile:
          type: string
        prompt:
          type: string
        releasePhase:
          type: string
        repairCount:
          format: int64
          type: integer
        repository:
          type: string
        reviewRunId:
          type: string
        revision:
          format: int64
          type: integer
        sessionId:
          type: string
        sourceBranch:
          type: string
        state:
          type: string
        target:
          type: string
        taskId:
          type: string
        updatedAt:
          format: date-time
          type: string
        worktreePath:
          type: string
      required:
      - taskId
      - payloadDigest
      - target
      - profile
      - repository
      - policyVersion
      - sessionId
      - cardNumber
      - worktreePath
      - sourceBranch
      - prompt
      - state
      - revision
      - repairCount
      - createdAt
      - updatedAt
      type: object
    DomainDCPV2Action:
      type: object
    DomainDCPV2Admission:
      type: object
    DomainDCPV2Command:
      type: object
    DomainDCPV2ExternalEvent:
      type: object
    DomainDCPV2Incident:
      type: object
    DomainDCPV2LifecycleProjection:
      properties:
        admissionId:
          type: string
        deployed:
          type: boolean
        detail:
          type: string
        error:
          type: boolean
        humanGate:
          type: boolean
        humanGateQuestion:
          type: string
        merged:
          type: boolean
        modelActive:
          type: boolean
        phase:
          enum:
          - worker_queued
          - worker_running
          - checks_waiting
          - review_queued
          - review_running
          - repair_queued
          - repair_running
          - arbiter_queued
          - arbiter_running
          - admission_waiting
          - readmission
          - release_waiting
          - merge_observing
          - release_verified
          - deployment_waiting
          - deployment_observing
          - human_gate
          - failed
          - merged
          - deployed
          type: string
        resultId:
          type: string
        revisionId:
          type: string
        role:
          type: string
        statusLabel:
          type: string
        workflowActive:
          type: boolean
      required:
      - phase
      - statusLabel
      - revisionId
      - modelActive
      - workflowActive
      - humanGate
      - error
      - merged
      - deployed
      type: object
    DomainDCPV2Result:
      type: object
    DomainDCPV2Revision:
      type: object
    DomainDCPV2Task:
      type: object
...
    ListDCPTaskEventsResponse:
      properties:
        events:
          items:
            $ref: '#/components/schemas/DCPTaskEventResponse'
          type: array
      required:
      - events
      type: object
    ListDCPTasksResponse:
      properties:
        tasks:
          items:
            $ref: '#/components/schemas/DCPTask'
          type: array
      required:
      - tasks
      type: object
...
    SubmitDCPTaskRequest:
      properties:
        approvedScope:
          $ref: '#/components/schemas/DCPApprovedScope'
        approvedTask:
          $ref: '#/components/schemas/DCPApprovedTask'
        idempotencyKey:
          maxLength: 128
          minLength: 1
          type: string
        target:
          enum:
          - dcp-lab
          type: string
      required:
      - idempotencyKey
      - target
      - approvedTask
      - approvedScope
      type: object
...